test migrated to typescript
