+++ remix-tests/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "include": ["src"],
   "compilerOptions": {
     /* Basic Options */
     "target": "es5",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017','ES2018' or 'ESNEXT'. */
@@ -16,7 +17,10 @@
     "paths": { "remix-tests": ["./"] },       /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
     "typeRoots": ["node_modules/@types"],                       /* List of folders to include type definitions from. */
     "esModuleInterop": true,                   /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
-
+    "types": [
+      "mocha",
+      "node"
+    ],
     /* Experimental Options */
     "experimentalDecorators": true,           /* Enables experimental support for ES7 decorators. */
   }
